refactor(code): cleaned folder structure
All the functions which fetch data -> under utils/ folder , renamed components -> pages and lib->
components
