+++ src/App.jsx
@@ -1,21 +1,18 @@
-import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
-import Login from './components/Login/Login'
-import { Routes,Route,Navigate } from 'react-router-dom'
-import Home from './components/Home/Home'
-import Callback from './components/Callback/Callback'
+import "./App.css";
+import Login from "./pages/Login/Login";
+import { Routes, Route, Navigate } from "react-router-dom";
+import Home from "./pages/Home/Home.jsx";
+import Callback from "./pages/Callback/Callback";
 
 function App() {
   return (
-     <Routes>
-      <Route path='/' element={<Home/>} />
-      <Route path='/login' element={<Login/>} />
-      <Route path='/callback' element={<Callback />} />
+    <Routes>
+      <Route path="/" element={<Home />} />
+      <Route path="/login" element={<Login />} />
+      <Route path="/callback" element={<Callback />} />
       <Route path="*" element={<Navigate to="/" replace />} />
-     </Routes>
-  )
+    </Routes>
+  );
 }
 
-export default App
+export default App;
