+++ src/components/button/Button.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import "./Button.css";
 
 const Button = ({
