+++ src/pages/Login/Login.jsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from "react";
 import "./Login.css";
-import Button from "../../lib/button/Button.jsx";
+import Button from "../../components/button/Button";
+import { checkLoginStatus } from "../../utils/checkLoginStatus";
 
 const Login = ({ onLogin }) => {
   const [isLoading, setIsLoading] = useState(true);
@@ -10,35 +11,13 @@ const Login = ({ onLogin }) => {
   const API_BASE_URL = import.meta.env.VITE_API_URL || "";
 
   useEffect(() => {
-    const checkLoginStatus = async () => {
-      try {
-        const token = localStorage.getItem("access_token");
-        if (!token) {
-          setIsLoading(false);
-          return;
-        }
-
-        const response = await fetch(`${API_BASE_URL}/auth/user`, {
-          credentials: "include",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${token}`,
-          },
-        });
-
-        if (response.ok) {
-          setLoggedIn(true);
-        } else {
-          setIsLoading(false);
-        }
-      } catch (err) {
-        setIsLoading(false);
-        setLoggedIn(false);
-        console.error("Error checking login status:", err);
-      }
+    const checkLogin = async () => {
+      const isLoggedIn = await checkLoginStatus(API_BASE_URL);
+      setLoggedIn(isLoggedIn);
+      setIsLoading(false);
     };
 
-    checkLoginStatus();
+    checkLogin();
   }, []);
 
   const handleGoogleLogin = async () => {
